move notebook constants to quartoNotebookSerializer, delete quartoConstants
Moves cell metadata types, language mappings, and cell boundary markers
from the shared quartoConstants.ts into the quartoNotebook module where
they are actually used. Removes quartoConstants.ts entirely.
